Cleanup DB models
- Rename `common.py` to `abc.py` and use it for base classes and mixins

- Move `GaiaBase` to `abc.py` and make it the mixin `CRUDMixin`

- Move the common models next to the classes they belong to (in app, gaia and system)

- Move the memory models next to the base class they belong to (in app, gaia and system)

- Use `gv.WarningLevel` instead of `ImportanceLevel`
