+++ src/ouranos/core/database/models/__init__.py
@@ -2,22 +2,19 @@ from ouranos.core.database.models.app import (
     anonymous_user, CalendarEvent, CommunicationChannel, FlashMessage,
     Permission, Role, Service, User, UserMixin)
 from ouranos.core.database.models.archives import (
-    ActuatorRecordArchive, HealthRecordArchive, SensorRecordArchive)
-from ouranos.core.database.models.common import ImportanceLevel
+    ActuatorRecordArchive, HealthRecordArchive, SensorDataRecordArchive)
 from ouranos.core.database.models.gaia import (
     ActuatorRecord, Ecosystem, EnvironmentParameter, Engine, GaiaWarning,
-    Hardware, HealthRecord, Lighting, Measure, Plant, SensorRecord)
-from ouranos.core.database.models.memory import SensorDbCache, SystemDbCache
-from ouranos.core.database.models.system import SystemRecord
+    Hardware, HealthRecord, Lighting, Measure, Plant, SensorDataCache, SensorDataRecord)
+from ouranos.core.database.models.system import SystemDataCache, SystemDataRecord
 
 
 __all__ = [
     "anonymous_user", "CalendarEvent", "CommunicationChannel",
     "FlashMessage", "Permission", "Role", "Service", "User", "UserMixin",
-    "ActuatorRecordArchive", "HealthRecordArchive", "SensorRecordArchive",
-    "ImportanceLevel",
+    "ActuatorRecordArchive", "HealthRecordArchive", "SensorDataRecordArchive",
     "ActuatorRecord", "Ecosystem", "EnvironmentParameter", "Engine",
     "GaiaWarning", "Hardware", "HealthRecord", "Lighting", "Measure", "Plant",
-    "SensorRecord",
-    "SystemRecord"
+    "SensorDataRecord",
+    "SystemDataRecord"
 ]
