+++ src/ouranos/web_server/validate/gaia/sensor.py
@@ -6,7 +6,7 @@ from pydantic import Field
 
 import gaia_validators as gv
 
-from ouranos.core.database.models.gaia import SensorRecord
+from ouranos.core.database.models.gaia import SensorDataRecord
 from ouranos.core.validate.base import BaseModel
 from ouranos.core.validate.utils import sqlalchemy_to_pydantic
 from ouranos.web_server.validate.gaia.hardware import HardwareInfo
@@ -38,7 +38,7 @@ class SensorSkeletonInfo(BaseModel):
 #   Current sensor data
 # ---------------------------------------------------------------------------
 _SensorRecordModel = sqlalchemy_to_pydantic(
-    SensorRecord,
+    SensorDataRecord,
     base=BaseModel,
     exclude=["id"]
 )
