+++ src/ouranos/web_server/validate/gaia/warning.py
@@ -2,13 +2,14 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from ouranos.core.database.models.common import ImportanceLevel
+import gaia_validators as gv
+
 from ouranos.core.validate.base import BaseModel
 
 
 class WarningInfo(BaseModel):
     id: int
-    level: ImportanceLevel
+    level: gv.WarningLevel
     title: str
     description: str
     created_on: datetime
