+++ tests/data/app.py
@@ -1,13 +1,13 @@
 from datetime import datetime, timedelta, timezone
 
-from ouranos.core.database.models.common import ImportanceLevel
+import gaia_validators as gv
 
 
 timestamp_now = datetime.now(timezone.utc)
 
 
 calendar_event = {
-    "level": ImportanceLevel.low,
+    "level": gv.WarningLevel.low,
     "title": "An event",
     "description": "That is not really important",
     "start_time": timestamp_now + timedelta(days=2),
